# IDGS901_API_Balones/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BalonEvo",
    "version": "v1"
  },
  "paths": {
...
    "/api/Tarjetas/{usuario}": {
      "post": {
        "tags": [
          "Tarjetas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tarjetas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Tarjetas"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Tarjetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tarjetas/{usuario}/{numTarjeta}": {
      "delete": {
        "tags": [
          "Tarjetas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "usuario",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "numTarjeta",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Clientes": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "rol": {
          "type": "string"
        },
        "estatus": {
          "type": "string"
        }
      }
    },
...
    "Tarjetas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombreTarjeta": {
          "type": "string"
        },
        "numTarjeta": {
          "type": "string"
        },
        "fechaVencimiento": {
          "type": "string"
        },
        "ccv": {
          "type": "string"
        },
        "usuario": {
          "$ref": "#/definitions/Clientes"
        }
      }
    },
...
  }
}